# Scala/Scala/projects/LearningTracker/conf/swagger.yaml
openapi: 3.0.3
info:
  title: Learning Tracker API
  version: 0.1.0
  description: Demo API for Programming in Scala language features.
servers:
  - url: http://localhost:9000
paths:
...
  /api/courses/{id}:
    get:
      summary: Get a course by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Course:
      type: object
      required: [id, title, level, status, lessons, tags, metadata, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        level:
          $ref: "#/components/schemas/Level"
        status:
          $ref: "#/components/schemas/Status"
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
        tags:
          type: array
          uniqueItems: true
          items:
            type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Lesson:
      type: object
      required: [id, title, minutes, tags]
      properties:
        id:
          type: string
        title:
          type: string
        minutes:
          type: integer
          minimum: 0
        tags:
          type: array
          items:
            type: string
...
    Level:
      type: string
      enum: [Beginner, Intermediate, Advanced]
    Status:
      type: string
      enum: [Planned, InProgress, Completed, Archived]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string